refactor(co): move spinlock to detail/
